Renames models file to phage.py as to not conflict with django models file. Also changes related scripts accordingly.
